last commit idk

+++ src/fnplot/values/FnPlotInt.java
@@ -88,7 +88,7 @@ public class FnPlotInt extends FnPlotValue<FnPlotInt> {
         if (arg.isInteger()) {
             return make((int)Math.pow(value, arg.intValue()));
         } else {
-            return make((int)Math.pow(value, arg.intValue()));
+            return make((int)Math.pow(value, arg.doubleValue()));
         }
     }
